requirements.txt
blis==0.7.11
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
confection==0.1.5
cycler==0.12.1
cymem==2.0.8
Cython==0.29.23
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
fonttools==4.54.1
gensim==4.1.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.1
mdurl==0.1.2
murmurhash==1.0.10
nltk==3.6.7
numpy==1.21.4
openpyxl==3.0.9
packaging==24.1
pandas==1.3.5
pillow==10.4.0
praw==7.5.0
prawcore==2.4.0
preshed==3.0.9
pydantic==1.10.9
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.8.1
scipy==1.7.3
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.5
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.7.1
update-checker==0.18.0
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
wrapt==1.16.0
